+++ doc/source/changelog.rst
@@ -14,6 +14,7 @@ MontePy Changelog
 
 **Features Added**
 
+* Implement _collection_ref to link objects to their NumberedObjectCollection parent (:issue:`867`).
 * Added checking for additional input after the ``data`` block, and raising a warning if it exists (:issue:`525`).
 * Allow multiple universe fills to accept 2D MNCP lattices (:issue:`719`).
 * Make ``LatticeType.RECTANGULAR`` and ``LatticeType.HEXAHEDRAL`` synonymous (:issue:`808`).